Add reusable UpdatedAgo component for refresh status
Introduces a new UpdatedAgo component to display last updated time and provide a refresh button. Refactors activity and logs tables to use this component, improving UI consistency and code reuse. Removes duplicate refresh logic from logs-table.
